requirements.txt
tensorflow==2.9.0
Django==3.2.18
Pillow==9.4.0
H5py==3.6.0
SciPy==1.7.3
openpyxl== 3.1.1
opencv-contrib-python
matplotlib
scikit-learn==1.0.2
PyQt5==5.12
protobuf==3.20.0
bs4
numpy